feat(both): add register function
when the user logs in, the backend will return 'exists': 'true' if the user exists, otherwise, 'exists': 'false'. If it doesn't exist, the program will be redirected to the register page, where the user is supposed to fill in some information. Note that the hash function of creating familyId is not implemented yet.